feat(user): added search by role and hire date, separated user service
